A month is a twelth of a year
This one can be debated. While it's obvious that one hour is
3600 seconds, one day is 24 hours and one week is 7 days, it's
more difficult to convince anyone that a month is 30 days.
It's probably easier to argue that 12 months make up one year,
but that doesn't really help since the number of seconds per
year can be different due to leap years. And leap seconds for
that matter.

The definition of a year in juration has always been 365 days,
or 31536000 seconds. The true number would probably be closer
to 31556926 if you take leap years and such into effect. But
the simplification of 365 days is okey if you know about
the limitation. The definition of a month was previously 30
days. With this commit, that number has been changed so that one
month is a twelth of a year, or 2628000 seconds.